8.5.45 (834633313)
We’re always making changes and improvements to Spotify. To make sure you don’t miss a thing, just keep your Updates turned on.

Bug fixes and improvements in this version include:
- Fixed stability issues

+++ SPStepScrollViewDataSource-Protocol.h
@@ -10,6 +10,6 @@
 
 @protocol SPStepScrollViewDataSource <NSObject>
 - (SPPageView *)stepScrollView:(SPStepScrollView *)arg1 pageViewAtRelativeIndex:(long long)arg2;
-- (long long)stepScrollView:(SPStepScrollView *)arg1 numberOfStepsInDirection:(int)arg2;
+- (long long)stepScrollView:(SPStepScrollView *)arg1 numberOfStepsInDirection:(long long)arg2;
 @end
